Refactored project and added CAN data
