+++ channels_to_motec/interpolate_channel.py
@@ -1,4 +1,3 @@
-import pandas as pd
 import numpy as np
