+++ channels_to_motec/mandrewLDWriter.py
