+++ example.py
@@ -1,15 +1,15 @@
-from mandrewLDWriter import File, Channel
-import pandas as pd
-import transform_channel
 import time
+import pandas as pd
+from channels_to_motec.mandrewLDWriter import File, Channel
+import channels_to_motec.transform_channel as tc
 
 # Load CSV files into dataframes
 steered_angle_dataframe = pd.read_csv('Data/Steered Angle Data/random_steering_data_from_grafana_10-13-2024.csv')
 throttle_position_dataframe = pd.read_csv('Data/Throttle Position Data/random_throttle_position_data_from_grafana_2024-10-22.csv')
 
 # Interpolate the channel to create regular time intervals, so we can use HZ in the MoTec converter I like
-processed_angle_series, angle_hz = transform_channel.process_steered_angle(steered_angle_dataframe['Time'], steered_angle_dataframe['data Analog Input #2'])
-processed_throttle_series, throttle_hz = transform_channel.process_throttle_position(throttle_position_dataframe['Time'], throttle_position_dataframe['data TPS'])
+processed_angle_series, angle_hz = tc.process_steered_angle(steered_angle_dataframe['Time'], steered_angle_dataframe['data Analog Input #2'])
+processed_throttle_series, throttle_hz = tc.process_throttle_position(throttle_position_dataframe['Time'], throttle_position_dataframe['data TPS'])
 
 # Create a File instance
 example_file = File()
